Add files via upload
This is the main code to load the data and calculate the peak statistics of the IZQC signal. The scriptIZQC does the job and calls the other functions which will be needed depending on the dataset that we are loading. 

The data for the paper can be found at:

https://mega.nz/folder/hugkzIiC#s3087pvVOsuHMJwJe3c-1g

If you experience any problem with the script or you get some errors please contact me at d.lopez@psych.pan.pl or david.lopez@psych.uw.edu.pl and I will help you through the process.